Move dialect-specific code to it's directory + implement against modifier
