+++ lib/sqlalchemy/sql/expression.py
@@ -66,7 +66,6 @@ __all__ = [
     "lambda_stmt",
     "literal",
     "literal_column",
-    "match",
     "not_",
     "null",
     "nulls_first",
@@ -123,7 +122,6 @@ from .elements import Grouping
 from .elements import Label
 from .elements import literal
 from .elements import literal_column
-from .elements import match
 from .elements import not_
 from .elements import Null
 from .elements import outparam
